Fix obsoletes
